src/test/java/kevins/fun/demo/endpoints/unit/controllers/UserControllerTest.java
package kevins.fun.demo.endpoints.unit.controllers;

import kevins.fun.demo.controller.UserController;
import kevins.fun.demo.endpoints.unit.util.JsonUtil;
import kevins.fun.demo.entity.User;
import kevins.fun.demo.repository.UserRepository;
import kevins.fun.demo.service.UserService;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.CoreMatchers.is;
import static org.mockito.BDDMockito.given;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;


@RunWith(SpringRunner.class)
@WebMvcTest(UserController.class)
class UserControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService service;

    @MockBean
    private UserRepository repository;

    @Test
    @DisplayName("Should Create A User When making POST request to endpoint - /core/users")
    public void createUser_whenPostMethod() throws Exception {

        User user = new User();
        user.setFirstName("Fan");
        user.setLastName("Kevin");

        given(service.create(user)).willReturn(user);

        mockMvc.perform(post("/core/users")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(JsonUtil.toJson(user)))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.firstName", is(user.getFirstName())))
                .andExpect(jsonPath("$.lastName", is(user.getLastName())));
    }
}
